backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
sentence-transformers>=2.3.0
torch==2.1.0
transformers==4.35.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
jupyter==1.0.0
ipykernel==6.26.0
matplotlib==3.8.2
seaborn==0.13.0
shap==0.43.0
openai>=1.0.0

